Modified version of ASCII Serial datasource to include additional parameters and options for serial devices.
This new datasource can be useful for devices that need an special
format and delimiters for the commands received from the SCADA isntance
and their responses.

It also includes the possibility for the SCADA to send a second and
third serial command after receiving a first or second response from the
serial device. This option is included to allow the capture of data from
a serial devices that need a handshake proccess to respond with the
requested data.