#1101 Improving parameter reading from env.properties
I added private variables to the ScadaConfig class that store configuration values.
If they are read once using the methods:
1. public Boolean getBoolean (String propertyName, boolean defaultValue)
2. public Long getLong (String propertyName, int defaultValue)
3. public Integer getInt (String propertyName, int defaultValue)
4. public String getProperty (String propertyName)

The values of the corresponding properties will be saved in the variables that will then be used when reading these values for specific configuration properties.

And I used ScadaConfig to read the configuration for running scripts in the ScriptExecutor class

I assume that in the future we will create a collection in ScadaConfig that will store a value but for now because of the ease of implementation and time I decided to use a simpler implementation.

fix #1101