#1107 Not saving the value of the NONE Alarm to the database
In EventService, the EventService changed the NONE type event record, which now does not write to the database.
UserEventDAO has changed the way user events are saved so that it does not attempt to save when that does not exist in the events table.
In .travis.yml - A line has been added to the Travis to build a separate docker for this task.

fix #1107