#1860 ADD User Profile Vue.js page behaviour, layot
Added all basic operations related with User Profiles like creation modification and deletion. User is able to change the properties and the watcher follow the user changes and if there is no changes save button is not visible. Attached that component into users details component.

Code cleanup and refactor is necessary.