#1902 Change the component UX based on the @grzesiekb suggestions.
Reduced the buttons visible on the component. Now the history and the state change is inside the menu box supported by specific headings.

Added new condition to check "disabled" in "analyzeInOrder" to disable controll buttons if they are blocked by some state. Also new prop like "hideControls" has been provided and it works in the same way as this disable option.