#1933 ADD NewPageNotification tag to old UI
When user opens a page that has its equivalent in ModernUI the notificaton is appering on the right corner of the page. It is possible to move using the "Move!" action, hide that notification using "cross button" or disable that popup by clicking "do not show anymore".

What is more the navigation to Modern Interface is realized by a specific navigation icon. It was added at the beginning of the navbar with a new icon that represent application. It is separated by the horizontal bar.

Removed ModernWatchList entry and also Alarms. To go to that pages user must use the new Icon.