#1934 Add HideHeader function to User
Now when User opens the new User Interface the new property is set to True to inform the application to work in embedded mode. When user use "Back to Old UI" button and open the WatchList page the controller disable that property and application work as normal. Added some catch clousure to handle the Event Flag errors when it is not visible.