#1951 Add ScadaVersion class and change UI
New Java class keeps the properties from "version" file. Some additional content is now possible to be rendered on the Header next to the Scada-LTS logo. We added "companyName" for which this application is configured. "poweredBy" for someone who manage the ScadaLTS application. Scada-LTS application version in a header and build number. What is more the GitHub commit reference is visible there also. So many many chages. The GitHub Action build script has been updated to pass this data to valid place inside a project.