#2012 Add and refactor Notification Center
All snackbar messages are now replaced by new Notification Component that handle the notification message and shows it in top right corner of the application. This mechanism supports more than one message that is displayed and is binded with global HTTP requests methods so when it receives a error message user is informed about that error.